+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 2.x branch
 ## 2.23 branch
 ### 2.23.0
+* the `ConsoleIntegrationTestTrait` provides the `$Command` property through the `__get()` magic method, and no longer
+  through the `setUp()` method.
 * the `ComponentTestCase` provides the `$Component` property through the `__get()` magic method, and no longer through
   the `setUp()` method.
 * the `HelperTestCase` provides the `$Helper` property through the `__get()` magic method, and no longer through the